refactor(forge): split the renderer by region, prune dead exports
draw.ts had grown to 1350 lines doing six unrelated jobs. It now owns only the
draw ORDER; each region is its own module along the seams that were already
marked by comment banners:

  backdrop.ts   the room, the fire, the furnace body and its empty bays
  chain.ts      the six station rows
  header.ts     readout, heat gauge, permanent multipliers, quench plate
  workbench.ts  anvil, work bar, hammer, cast ingots
  ephemera.ts   floating figures, thrown ingots, the milestone stamp
  overlays.ts   seal, haul, forge mark, quench
  common.ts     fitSize / chevron / chains, shared by the above

Also removed exports nothing calls — roundRect, FONT_LABEL, FONT_NUM, GLOW_RED,
unitsToMicro, microToUnits, scaleRat, bigMin, bigMax, ZERO, toNextDouble,
wipe — and made glowSprite, platePattern, digits and groupDigits module-private.
`noUnusedLocals` does not catch an unused export, so these accumulate silently.

Pure reorganisation: the production bundle is byte-identical at 62.00 kB, all
49 tests pass, and the rendered frame was compared against the pre-split build.
